Patch_tool/requirement.txt
fastapi
uvicorn
psycopg2-binary
requests
psutil
